feat(templates): improve SignalR implementation in Boilerplate (#8557) (#8643)
