+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Client/Boilerplate.Client.Core/Routes.razor.cs
@@ -2,73 +2,6 @@
 
 public partial class Routes
 {
-    [AutoInject] IJSRuntime jsRuntime = default!;
-    [AutoInject] IBitDeviceCoordinator bitDeviceCoordinator = default!;
-    [AutoInject] IStorageService storageService = default!;
-    [AutoInject] CultureInfoManager cultureInfoManager = default!;
-
-    protected override async Task OnInitializedAsync()
-    {
-        if (AppPlatform.IsBlazorHybrid)
-        {
-            if (CultureInfoManager.MultilingualEnabled)
-            {
-                cultureInfoManager.SetCurrentCulture(await storageService.GetItem("Culture") ?? // 1- User settings
-                                                     CultureInfo.CurrentUICulture.Name); // 2- OS settings
-            }
-
-            await SetupBodyClasses();
-        }
-
-        await base.OnInitializedAsync();
-    }
-
-    private async Task SetupBodyClasses()
-    {
-        var cssClasses = new List<string> { };
-
-        if (AppPlatform.IsWindows)
-        {
-            cssClasses.Add("bit-windows");
-        }
-        else if (AppPlatform.IsMacOS)
-        {
-            cssClasses.Add("bit-macos");
-        }
-        else if (AppPlatform.IsIOS)
-        {
-            cssClasses.Add("bit-ios");
-        }
-        else if (AppPlatform.IsAndroid)
-        {
-            cssClasses.Add("bit-android");
-        }
-
-        var cssVariables = new Dictionary<string, string>();
-        var statusBarHeight = bitDeviceCoordinator.GetStatusBarHeight();
-
-        if (AppPlatform.IsMacOS is false)
-        {
-            //For iOS this is handled in css using safe-area env() variables
-            //For Android there's an issue with keyboard in fullscreen mode. more info: https://github.com/bitfoundation/bitplatform/issues/5626
-            //For Windows there's an issue with TitleBar. more info: https://github.com/bitfoundation/bitplatform/issues/5695
-            statusBarHeight = 0;
-        }
-
-        cssVariables.Add("--bit-status-bar-height", $"{statusBarHeight.ToString("F3", CultureInfo.InvariantCulture)}px");
-        await jsRuntime.ApplyBodyElementClasses(cssClasses, cssVariables);
-    }
-
-    protected async override Task OnAfterRenderAsync(bool firstRender)
-    {
-        await base.OnAfterRenderAsync(firstRender);
-
-        if (firstRender && AppPlatform.IsBlazorHybrid is false)
-        {
-            AppPlatform.OSDescription = await jsRuntime.GetBrowserPlatform();
-        }
-    }
-
     [AutoInject] NavigationManager? navigationManager { set => universalLinksNavigationManager = value; get => universalLinksNavigationManager; }
     public static NavigationManager? universalLinksNavigationManager;
