+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Shared/appsettings.json
@@ -2,14 +2,14 @@
     "Logging": {
         "LogLevel": {
             "Default": "Warning",
-            "Boilerplate.Client.Core.Components.AuthenticationStateLogger": "Information",
+            "Boilerplate.Client.Core.Services.AuthenticationManager": "Information",
             "Microsoft.EntityFrameworkCore.Database.Command": "Information"
         },
         //#if (appCenter == true)
         "AppCenterLoggerProvider": {
             "LogLevel": {
                 "Default": "Warning",
-                "Boilerplate.Client.Core.Components.AuthenticationStateLogger": "Information",
+                "Boilerplate.Client.Core.Services.AuthenticationManager": "Information",
                 "Microsoft.EntityFrameworkCore.Database.Command": "Information"
             }
         },
@@ -18,7 +18,7 @@
         "ApplicationInsights": {
             "LogLevel": {
                 "Default": "Warning",
-                "Boilerplate.Client.Core.Components.AuthenticationStateLogger": "Information",
+                "Boilerplate.Client.Core.Services.AuthenticationManager": "Information",
                 "Microsoft.EntityFrameworkCore.Database.Command": "Information"
             }
         }
